requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Computer vision
ultralytics>=8.0.0
opencv-python>=4.8.0

# Configuration and logging
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0

# Data and metrics
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Optional (for SAM2 segmentation)
# segment-anything-2
